Backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
openai==0.27.0
python-dotenv==1.0.0